Return on hacktoberfest PR and move code
Adjusted the hacktoberfest handler to return a bool if the PR was
determined to be a 'spam' PR so that the process can exit early
without sending the DCO, sign-off, and other messages to the PR.

Also, completed the file rename and move to keep the variables
and tests in their appropriate places.
