+++ handler/pullrequest_handler_test.go
@@ -210,75 +210,6 @@ func Test_hasAnonymousSign(t *testing.T) {
 	}
 }
 
-func Test_onlyMarkdownFiles(t *testing.T) {
-	mdFileName1 := "readme.md"
-	mdFileName2 := "README.MD"
-	nonMDFileName := "main.go"
-
-	var testCommits = []struct {
-		files    []*github.CommitFile
-		expected bool
-	}{
-		{
-			files: []*github.CommitFile{
-				&github.CommitFile{
-					Filename: &mdFileName1,
-				},
-			},
-			expected: true,
-		},
-		{
-			files: []*github.CommitFile{
-				&github.CommitFile{
-					Filename: &mdFileName2,
-				},
-			},
-			expected: true,
-		},
-		{
-			files: []*github.CommitFile{
-				&github.CommitFile{
-					Filename: &mdFileName1,
-				},
-				&github.CommitFile{
-					Filename: &mdFileName2,
-				},
-			},
-			expected: true,
-		},
-		{
-			files: []*github.CommitFile{
-				&github.CommitFile{
-					Filename: &nonMDFileName,
-				},
-			},
-			expected: false,
-		},
-		{
-
-			files: []*github.CommitFile{
-				&github.CommitFile{
-					Filename: &mdFileName1,
-				},
-				&github.CommitFile{
-					Filename: &mdFileName2,
-				},
-				&github.CommitFile{
-					Filename: &nonMDFileName,
-				},
-			},
-			expected: false,
-		},
-	}
-
-	for _, test := range testCommits {
-		onlyMD := onlyMarkdownFiles(test.files)
-		if onlyMD != test.expected {
-			t.Errorf("Only markdown files - wanted %t, found %t", test.expected, onlyMD)
-		}
-	}
-}
-
 func stringPtr(s string) *string {
 	return &s
 }
