+++ main.go
@@ -92,7 +92,10 @@ func handleEvent(eventType string, bytesIn []byte, config config.Config) error {
 		if req.Action != handler.ClosedConstant {
 			contributingURL := getContributingURL(derekConfig.ContributingURL, req.Repository.Owner.Login, req.Repository.Name)
 			if handler.EnabledFeature(hacktoberfest, derekConfig) {
-				handler.HandleHacktoberfestPR(req, contributingURL, config)
+				isSpamPR, _ := handler.HandleHacktoberfestPR(req, contributingURL, config)
+				if isSpamPR {
+					return nil
+				}
 			}
 			if handler.EnabledFeature(dcoCheck, derekConfig) {
 				handler.HandlePullRequest(req, contributingURL, config)
